added selectPDFTextElement custom cypress command
