+++ polaris-ui/.gitignore
@@ -31,4 +31,4 @@ yarn-error.log*
 /cypress/screenshots
 /cypress/videos
 /coverage-merged
-/.vscode
+/.vscode/settings.json
